hw2/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
datamodel-code-generator
python-jose[cryptography]
passlib[bcrypt]
